fix: Move `StatusBadge` to `CustomCodes`
